TextMapper: Fix returning reference to temporary
TextMapper::getText() returns a reference.
Commit d98e41849f1b84d86ae394758f1bc5defdfbf05d makes it
so that missing entries get the string `"[" + key + "]"`.

This string lives in the getText function.
Returning a reference to it results in a dangling reference
after the function is finished.

This is undefined behavior, and causes a crash on my system.

This commit fixes this problem by returning a new string instead of
just a reference.